presidio-analyzer/pyproject.toml
[build-system]
build-backend = "poetry.core.masonry.api"
requires = ["poetry-core"]

[project]
name = "presidio_analyzer"
version = "2.2.360"
description = "Presidio Analyzer package"
authors = [{name = "Presidio", email = "presidio@microsoft.com"}]
license = "MIT"
classifiers = [
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]
keywords = ["presidio_analyzer"]
urls = {Homepage = "https://github.com/Microsoft/presidio"}
readme = "README.md"
include = ["conf/*",]

requires-python = ">=3.9,<3.14"
dependencies = [
    "spacy (>=3.4.4, !=3.7.0)",
    "regex",
    "tldextract",
    "pyyaml",
    "phonenumbers (>=8.12,<10.0.0)",
    "lingua (>=4.15.0,<5.0.0)",
    "langdetect (>=1.0.9,<2.0.0)",
    "sortedcontainers (>=2.4.0,<3.0.0)",
    "psutil (>=7.1.0,<8.0.0)"
]

[project.optional-dependencies]
server = [
    "flask (>=1.1)",
    "gunicorn"
]
transformers = [
    "transformers",
    "accelerate",
    "huggingface_hub",
    "spacy_huggingface_pipelines"]

stanza = [
    "stanza (>=1.10.1,<2.0.0)",
]
azure-ai-language = [
    "azure-ai-textanalytics",
    "azure-core",
]
ahds = [
    "azure-identity (>=1.23.0,<2.0.0)",
    "azure-health-deidentification (>=1.1.0b1,<2.0.0)"
]
gliner = [
    "transformers",
    "huggingface_hub",
    "gliner (>=0.2.13,<1.0.0) ; python_version >= '3.10'",
    "onnxruntime (>=1.19) ; python_version >= '3.10'"
]

[tool.poetry.group.dev.dependencies]
pip = "*"
ruff = "*"
pytest = "*"
pytest-mock = "*"
python-dotenv = "*"
pre_commit = "*"
